テストケース追加とメソッド分割

+++ CalorieCalculator/AmountOfIntake.cs
@@ -1,9 +1,23 @@
 ﻿namespace CalorieCalculator
 {
+    /// <summary>
+    /// 摂取量
+    /// </summary>
     public record AmountOfIntake
     {
-        public double Protein { get; init; }
-        public double Fat { get; init; }
-        public double Carbohydrate { get; init; }
+        /// <summary>
+        /// タンパク質
+        /// </summary>
+        public decimal Protein { get; init; }
+
+        /// <summary>
+        /// 脂肪
+        /// </summary>
+        public decimal Fat { get; init; }
+
+        /// <summary>
+        /// 炭水化物
+        /// </summary>
+        public decimal Carbohydrate { get; init; }
     }
 }
